Fix: Arrumando 80% de tudo 18/06 - 01:16

+++ database/migrations/2024_06_13_193251_criar_tabela_skill_vacancy.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('vacancy_id');
             $table->string('name');
-            $table->unsignedTinyInteger('level'); // Nível da habilidade (1 a 5)
+            $table->integer('level'); // Nível da habilidade (1 a 5)
             $table->timestamps();
         });
     }
